First draft of the php-wasm-browser docs
